requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.13.3
pydantic==2.6.3
pydantic_core==2.16.3
pyserial==3.5
sniffio==1.3.1
tk==0.1.0
tqdm==4.66.2
typing_extensions==4.10.0
